Created a .gitignore file first, preparation
I forgot to initialize a git repository when starting this project,
I have not tracked my progress so far.
I am starting with a .gitignore file because I want to filter the files
I do not want to add. I plan to use the command for adding
every file to the staging area.
I'm not sure if what I am doing is correct.