Adding every file to the main branch
I forgot to do this in the earlier phase, I did not think of using
git when I started the project, I thought that I should just make
a new project for the actual project, but then it would take too much
time, so I am using this project as a base for my actual project.
I am adding every file of this project to the master branch of the
repository, I don't know how to use branches. So I am sticking with
a single branch for now.