Updated the Nav links for pages
I started with fixing the preload file, and divided the functions
inside of a single export to multiple exports to the Main World.
I did this because I don't want long chains for accessing a
function, when the function chain can be shorten.
After fixing the preload file, I have to update the references to
the functions that I have changed inside of the preload.
Change the preload then update the references.
I changed the references, removed the "functions" export
from the chain, and replaced some function names, and removed some
functions and objects for testing out the Inter Process Communication.