Comments to show faults of register functions
Made comments in order to explain the missing parts of the register
fucntions, I did not create any sort of filter-like functionality
to prevent entry of invalid input for the item.
Will create the filters after creating the base fucntionality of the
program, the basic fucntions of the pages, for now I do not want to
focus too much attention in one page, I want to create the basic
structure of the program first.