Created the foundation for the purchase feature
1.Added html elements for receiving inputs, for making transactions,
the user needs to enter fields for identifying and modifying the items.
for purchasing, the user needs to enter information, textfields
and a button has been added to the purchase page.
2.Added functions for the purchase functionality, the process for
purchasing needs functions in order to keep code organize.
I have created the names and bodies, but I haven't written the
implementation of the code inside the function.
The functions have names but no code inside so even if called nothing
happens, I have not yet come up on how to actually do it yet. will
make steady progress later.
I added a comment to outline the process of the purchase.
3.Added a css file for the html page, I still don't plan on actually
adding any css yet, in preparation for future changes.