Added data commication between purchase and main
1.Created a way for the render process, preload, and main process
to communicate with each other, added an invoke call in the preload
file and created a handle for the event inside the index.js for
the main process.
2.Changed the data type of the Item Inventory, changed the data type
from array of item object to map of item object, though the item is
an prototype
3.Added functionality to the function responsible for taking the name
of the item the user wants to buy, the function will now retrieve input
and check if it does exist inside of the inventory, the result is of
a boolean value to indicate if it does exist.
4.Created a css file for the register-page.html, in order to prepare
for the future.
5.Used an asynchronous function with the process for taking the name of
the item to purchase because I need to wait for the respond of the main
process.