Started the body of the purchase functions
1.I modified the fucntion for taking the name of the item to purchase.
Created a return statement to return an object that holds fields
to describe the result, I used an object because I needed a boolean
variable for the controlling the flow of the program
2.Created basic verification guards on the functions for taking the
quantity and money for purchase. In order to prevent invalid input,
the guard pattern is implemented along with helper functions to filter
out unnecessary values.
3.Created helper functions for the main purchase functions, for
taking values such as: item name, quantity to purchase, money for
purchase.
4.Most of the functions are asynchronous because I need the functions
to wait for the execution of the fucntions that are fetching data from
the main process, I did not intend to execute the functions
concurrently, I intend to follow a line-by-line execution of the process
for purchasing.
Which is odd since I am using asynchronous functions for the
functionality of a synchronous flow, but I cannot use the await
keyword outside of an asynchronous function because it will stop the
main execution thread, so an asynchronous function is needed in order
to stop and wait for an execution.
Which is why I used multiple asynchronous functions in order to
utilize the await keyword.