Added an implementation for taking quantity to buy
1.Created the body of the function for taking the value that
represents the quantity to be purchased regarding the selected item
by the user.

    The function filters the contents to check if the value is valid
    then the value is passed to the main process to check
    the inventory in order to find out if it exist inside the inventory,
        if it exist
            proceed to next step,
        if not then
            send data to signal that item is not available,
            and end the function
    the next step is to check whether or not the supply can suffice for
    the purchase demand,
    if yes
        then proceed to send data for the next function for purchasing,
        and end the function
    then if not
        send data to signal no available supply for the demand
        and end the function

The function implements the guard-pattern in order to reduce nesting.
I think something is not right about the usage of the asynchronous
function and function calls.

This function is subject to future changes: I have not properly tested
the behaviour of the function.

2.Added a comment to the function, the for taking the quantity to be
purchased, the comment runs down the process, the original intended
purpose of the

3.Added a new handle to the main process, the handle will send the
data requested from the renderer process of the purchase panel.
The data to be returned is incased inside of an object, with a field
to indicate if the item is available given the amount being requested,
a field for the price per piece of the item.

4.Added a new function in the preload file in order to receive item
data from the main process in order to verify the item quantity to
purchase.

5.Used asynchronous functions not because of necessity, but because I
don't know how to properly handle Inter Process Communication