Started on purchase function for taking money
1. Started the body of the function for taking the money for
the completion of a purchase, the function is not finished and subject
to change soon because I am not yet done with it.

For now there are filters in order to identify invalid input values
because it would be problematic during testing if not implemented early
on

2. Created a CLass to represent the status of the function for taking
the money for purchase, whether or not the execution of the function
is smooth or there are problems, if there are no problems then the
fields of the object of the class would be modified,
in the case of a problem, the  default value would be enough
by default the class fields are set for a scenario where a problem
in the execution is present.