Changed the take-money function to fix bugs
Changed the take-money function to fix bugs and modified the preload
and index files to support the functionality of the
purchase-functionality.js.

!!!No proper return type for the take-money function

Created functions to facilitate the
communication and processing of data between main and render process.

1.Modified the take-money function for purchase process.

Added a new argument for the function to represent the name of the
item for purchase, the data will be used in the function for
creating transaction details.

Created a return value for the proper completion of the method, the
function will return a value that represents the data about the
transaction.

2.Fixed a bug in line 65, reversed the boolean value for the if
statement, previously causing the program to stop execution.

3. Created comments to describe the functionality of the classes
that represents the status of the take-money function.

Removed unnecesssary comments, there is no connection between the
current code and the comment, so removed it.

4.Created classes to hold the information about the status of taking
the money the user have entered.

The class TakeUserMoneyRespond and TransactionRecords, holds
information about the status of the function, the success and data
that the take-money function must return.

5. Exposed a new object/function in the preload file to modify the
inventory, the function is for subtracting the amount of an item
inside of the inventory, this is for when a purchase is made.

6. Created a new handle in the main process, added a new handle to
process the request from the render process to modify the inventory
for transaction purposes.

Added comments to describe chunks of the code of the main process.