Fixed 4 bugs on take-money function
tested the take-money function.
    I did not test the take-money function before so had to test it out
    during runtime and the problems appeared only there.

removed unnecessary comments.
    I made a lot of comments to de-activate code, so that I can reuse
    the code later.

Found bugs.
    did not solved it, but temporarily made the code work, not sure
    if the solution will last. The teest were quick and not thorough.

    purchase-page-functionality.js
        line 59, solved on line 56
        line 171, solved on line 181

    I cannot specify the line, but I encountered problems with
    the number data type and the string data type.

        I kept getting errors, it is not possible to compare a sting
        with a number.

        I learned a lot about how to use the debugger in electron and
        the internal dev tool of the electron browser window, that the
        main process and render process debugger can be switched by
        going into the fucntion call. From the render process, going
        into function calls that are connected to the main process
        through the preload will result in opening up the debugger in
        visual studio code for the main process debugging.

I used the get and set bindings for the class wrong, an error message
appeared on the console during runtime about a typeerror, fized by
googling the answer from:
"https://www.javascripttutorial.net/es6/javascript-getters-and-setters/".

Learned about the syntax before continuing with testing the code,
I thought that the class in java and javascript are the same,
which led me to believe that the behaviour of classes in javascript
would be the same in java.

So I got an error.