Major changes to Site Navigation and new files
Removed the event receivers in the main process, replaced it with
a single event handler, there were multiple on functions for Site
Navigation purposes, now there is just one on function for handling
Site Navigation.

I used multiple channels for the site navigation, I trimmed those
channels to a single channel. The purpose of the multiple channels was
to reload the browser window to a new page, I reduced it by using the
event data that can be passed from render process to main process to
signal what site to load to the browser window.

Created new files for the Transactions History page, an html, css, and
js file.