Added necessary dependency folders to repo
The folders with the dependency details have been created long
ago when the project started development but I added them to
the .gitignore.

I removed the names of the folders and files from the .gitignore
because I have to recraete the project in another computer, the
computer in the computer laboratory.

Problem, is that I don't know what version of electron I used for the
project and what are the dependency of the whole project. So in order
to fix the problem.

By dependency, I meant the settings for the workspace in visual studio
code and packages used for the electron project.